renamed a few tags and attributes
TITLEHEAD -> TITLE: the usage of this tag has changed. It used to be typographical, but now we'd just like to mark up all uses of the play name.
FONTGROUP -> FONT: this tag is needlessly verbose. Its attribute @n has also been changed to @size for clarity.
LINGROUP/@form -> @t: this is more consistent with other tagging.
RULE/@n, SPACE/@n, INDENT/@n -> @l: these are lengths, and are different enough from other uses of @n to warrant a different name.