Start working on a RULE tag validator. It's a biggie.
Separate out text and newline tokens. They're both treated as TextNode
elements in the DOM. This is necessary to get truer line numbers,
especially when a single text node runs over more than one line. It's
rare, but it happens.

Add a getParsedLine() method to DOM, to get all the nodes from one
line.

Add isWs() to TextNodes - true if the node is only whitespace.

Finally, add a rule validator which makes sure that RULE tags are on a
line with no other content. Whitespace and comments are allowed.

The normalizer transformer will combine adjacent
text nodes, so there's no worry there.

Work on #11